successfully display descriptions
